Fixed tests, added file modified time checking for remote files

+++ grib2.py
@@ -15,7 +15,9 @@ def download(data_dir):
     args:
         data_dir Directory to store data files
     """
-    import urllib, re, os, sys
+    import urllib, re, os, sys, time
+    from datetime import datetime
+    from dateutil import tz
 
     # Loop over directories that have forecast data files
     for dir in ['VP.001-003','VP.004-007']: # loop over remote directories
@@ -40,7 +42,16 @@ def download(data_dir):
             if line.find(".bin") != -1:
 
                 # Split line to get date and filename
-                month, day, time, filename = re.split("\s+", line)[5:9]
+                month, day, rtime, filename = re.split("\s+", line)[5:9]
+
+                # Convert date to local timezone
+                from_zone = tz.tzutc()
+                to_zone = tz.tzlocal()
+                today = datetime.today()
+                year = today.strftime('%Y')
+                utc = datetime.strptime('{0}/{1} {2} @ {3}'.format(month, day, year, rtime), '%b/%d %Y @ %H:%M')
+                utc = utc.replace(tzinfo = from_zone)
+                remote_time = time.mktime(utc.astimezone(to_zone).timetuple())
 
                 # Split filename to get noaa param name
                 param = filename.split('.')[1]
@@ -48,12 +59,21 @@ def download(data_dir):
                 # Only download files if we are interested in this parameter
                 if param in noaa_params:
 
-                    info("%s last modified %s/%s @ %s" % (filename, month, day, time))
+                    info("%s last modified %s/%s @ %s" % (filename, month, day, rtime))
 
-                    # Use wget -N to only download if last modified date has changed
                     remote_path = "%s/%s/%s" % (base_url, dir, filename)
                     local_path = "{0}/{1}/{2}".format(data_dir, dir, filename)
-                    urllib.urlretrieve (remote_path, local_path)
+                    local_time = os.stat(local_path).st_mtime
+
+                    # If file does not exist or the local file is older than the remote file, download
+                    if not os.path.exists(local_path) or local_time < remote_time:
+                        info('Downloading remote file')
+                        urllib.urlretrieve (remote_path, local_path)
+                        os.utime(local_path, (remote_time, remote_time))
+                    # Otherwise, just log some information
+                    else:
+                        info('Local file modified: {0} - Remote file modified: {1}'.format(local_time, remote_time))
+                        info('Local file is up-to-date, skipping download')
                     
     # Cube data files if any were downloaded - TODO check to see if files were actually downloaded
     cmd = "{degrib} {data_dir}/VP.001-003/*.bin {data_dir}/VP.004-007/*.bin -Data -Index {data_dir}/all.ind -out {data_dir}/all.dat".format(
